requirements.txt
# Python project requirements
streamlit==1.32.0
python-dotenv==1.0.0
openai==1.12.0
scikit-learn==1.4.0
numpy==1.26.0
pandas==2.1.0
httpx==0.27.0
python-dotenv==1.0.0
